+++ frontend/packages/ui/src/blocks/DividerBlock.vue
@@ -3,5 +3,5 @@
 </script>
 
 <template>
-  <hr />
+  <hr class="border-line" />
 </template>